fixup! EZP-31113: AssignSectionToSubtree event is not handled by search and HTTP cache
